+++ CHANGELOG.md
@@ -22,6 +22,8 @@
 - `examples`
   - [#850](https://github.com/wix-incubator/rich-content/pull/850) fix performance
   - [#850](https://github.com/wix-incubator/rich-content/pull/850) save content to local storage
+- `wrapper`
+  - [#852](https://github.com/wix-incubator/rich-content/pull/852) strategies create configs & not functions
 
 </details>
 <hr/>